requirements.txt
networkx==2.6.3
matplotlib==3.4.3
pygraphviz==1.7
pandas==1.3.3
python-igraph==0.9.7
scipy==1.7.3
tqdm==4.64.0
ogb==1.3.3
